refactor: 도메인 및 패키지 구조 리팩토링
- Theme 및 ReservationTime 관련 클래스/인터페이스를 각자의 도메인별 패키지로 이동
- Status enum을 ReservationStatus로 변경
- 불필요한 static 변수 제거 및 final 필드 추가
- 관련된 import 경로 수정
