+++ src/main/java/roomescape/global/jwt/TokenInfo.java
@@ -3,8 +3,9 @@ package roomescape.global.jwt;
 import roomescape.member.domain.Role;
 
 public class TokenInfo {
-    private Long id;
-    private Role role;
+
+    private final Long id;
+    private final Role role;
 
     public TokenInfo(Long id, Role role) {
         this.id = id;
