+++ src/main/java/roomescape/reservation/domain/Reservation.java
@@ -11,6 +11,8 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import java.time.LocalDate;
 import roomescape.member.domain.Member;
+import roomescape.reservation.time.domain.ReservationTime;
+import roomescape.theme.domain.Theme;
 
 @Entity
 public class Reservation {
@@ -36,13 +38,13 @@ public class Reservation {
 
     @Column(nullable = false)
     @Enumerated(value = EnumType.STRING)
-    private Status status;
+    private ReservationStatus status;
 
     public Reservation() {
     }
 
     public Reservation(final Long id, final Member member, final Theme theme, final LocalDate date,
-                       final ReservationTime reservationTime, final Status status) {
+                       final ReservationTime reservationTime, final ReservationStatus status) {
         this.id = id;
         this.member = member;
         this.theme = theme;
@@ -53,16 +55,16 @@ public class Reservation {
 
     public static Reservation createReserved(final Member member, final Theme theme, final LocalDate date,
                                              final ReservationTime reservationTime) {
-        return new Reservation(null, member, theme, date, reservationTime, Status.RESERVED);
+        return new Reservation(null, member, theme, date, reservationTime, ReservationStatus.RESERVED);
     }
 
     public static Reservation createWaiting(final Member member, final Theme theme, final LocalDate date,
                                             final ReservationTime reservationTime) {
-        return new Reservation(null, member, theme, date, reservationTime, Status.WAITING);
+        return new Reservation(null, member, theme, date, reservationTime, ReservationStatus.WAITING);
     }
 
     public void acceptStatus() {
-        status = Status.RESERVED;
+        status = ReservationStatus.RESERVED;
     }
 
     public Long getId() {
@@ -85,7 +87,7 @@ public class Reservation {
         return theme;
     }
 
-    public Status getStatus() {
+    public ReservationStatus getStatus() {
         return status;
     }
 }
