+++ src/main/java/roomescape/reservation/domain/ReservationStatus.java
@@ -1,12 +1,12 @@
 package roomescape.reservation.domain;
 
-public enum Status {
+public enum ReservationStatus {
     RESERVED("예약"),
     WAITING("대기");
 
     private final String status;
 
-    Status(final String status) {
+    ReservationStatus(final String status) {
         this.status = status;
     }
