+++ src/main/java/roomescape/reservation/presentation/controller/AdminReservationController.java
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 import roomescape.global.auth.Auth;
 import roomescape.member.domain.Role;
-import roomescape.reservation.application.service.ReservationService;
+import roomescape.reservation.application.ReservationService;
 import roomescape.reservation.presentation.dto.AdminReservationRequest;
 import roomescape.reservation.presentation.dto.ReservationResponse;
