+++ src/main/java/roomescape/reservation/time/domain/ReservationTime.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.domain;
+package roomescape.reservation.time.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
