+++ src/main/java/roomescape/theme/application/ThemeRepository.java
@@ -1,10 +1,10 @@
-package roomescape.reservation.application.repository;
+package roomescape.theme.application;
 
 import java.util.List;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.ListCrudRepository;
 import org.springframework.stereotype.Repository;
-import roomescape.reservation.domain.Theme;
+import roomescape.theme.domain.Theme;
 
 @Repository
 public interface ThemeRepository extends ListCrudRepository<Theme, Long> {
