+++ src/main/java/roomescape/theme/domain/Theme.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.domain;
+package roomescape.theme.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
