+++ src/test/java/roomescape/reservation/domain/ReservationTimeTest.java
@@ -9,6 +9,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.test.context.ActiveProfiles;
+import roomescape.reservation.time.domain.ReservationTime;
 
 @ActiveProfiles("test")
 @DataJpaTest
