+++ src/test/java/roomescape/reservation/domain/ThemeTest.java
@@ -8,6 +8,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.test.context.ActiveProfiles;
+import roomescape.theme.domain.Theme;
 
 @ActiveProfiles("test")
 @DataJpaTest
